refactor(observability): re-home telemetry + monitoring onto aragora.observability (VAL-P4A-010/011)
Move the telemetry collector + research-events originals and the monitoring
SimpleObserver into aragora/observability/, leaving DeprecationWarning re-export
shims at the legacy aragora.telemetry.* and aragora.monitoring.* paths. No caller
migration (telemetry floor stays 0; the single monitoring consumer stays on the
shim so the ceiling stays 1). Generated METRICS.md / module_tiers.yaml are not
regenerated here (path-frozen by open PRs); drift deferred to the phase merge-train.
